requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-appconf==1.0.4
django-imagekit==4.0.2
django-indexer==0.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-paging==0.2.5
django-rest-swagger==2.2.0
django-sentry==1.13.5
django-templatetag-sugar==1.0
djangorestframework==3.12.2
dnspython==1.16.0
docutils==0.16
eventlet==0.30.1
greenlet==1.0.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
pilkit==2.0
Pillow==8.1.0
psycopg2-binary==2.8.6
python-daemon==2.2.4
pytils==0.3
pytz==2021.1
raven==6.10.0
requests==2.25.1
sentry-sdk==0.19.5
simplejson==3.17.2
six==1.15.0
South==1.0.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
